+++ emp-ot/ferret/constants.h
@@ -5,13 +5,13 @@ namespace emp {
 static std::string PRE_OT_DATA_REG_SEND_FILE = "./data/pre_ot_data_reg_send";
 static std::string PRE_OT_DATA_REG_RECV_FILE = "./data/pre_ot_data_reg_recv";
 
-const static int N_REG = 10608640;
-const static int T_REG = 1295;
-const static int K_REG = 589824;
-const static int BIN_SZ_REG = 13;
-const static int N_PRE_REG = 649728;
-const static int T_PRE_REG = 1269;
-const static int K_PRE_REG = 36288;
-const static int BIN_SZ_PRE_REG = 9;
+const static int64_t N_REG = 10608640;
+const static int64_t T_REG = 1295;
+const static int64_t K_REG = 589824;
+const static int64_t BIN_SZ_REG = 13;
+const static int64_t N_PRE_REG = 649728;
+const static int64_t T_PRE_REG = 1269;
+const static int64_t K_PRE_REG = 36288;
+const static int64_t BIN_SZ_PRE_REG = 9;
 }//namespace
 #endif //EMP_FERRET_CONSTANTS_H__
