+++ emp-ot/ot.h
@@ -6,8 +6,8 @@ namespace emp {
 
 template<typename T>
 class OT { public:
-	virtual void send(const block* data0, const block* data1, int length) = 0;
-	virtual void recv(block* data, const bool* b, int length)  = 0;
+	virtual void send(const block* data0, const block* data1, int64_t length) = 0;
+	virtual void recv(block* data, const bool* b, int64_t length)  = 0;
 	virtual ~OT() {
 	}
 };
